samples/python/publish-websockets/requirements.txt
numpy~=2.3.0
paho-mqtt>=2.0.0
PySocks
